First steps in migrating checkout to object-based workflow. Introduces `dvc.objects.diff` that can be used in other applications (e.g. in `dvc diff`) later. Allows us to stage existing target first(multithreaded, benefiting from all current and future improvements in `dvc.objects.stage`), then compare the resulting object with the end result that we are trying to achieve and use that diff to only apply actual changes.

`diff()` has some pre-requisites for supporting granular `Tree`s in the future, but for the most part this is trying to carry over legacy things, that will be revisited in the followups.

Pre-requisite for #6594

Todo:
- [x] investigate checkpoints test failing
- [x] check benchmarks