@efiop 2021-09-21
benchmarks so far:

master:
```
---------------------------------------------
test_checkout-checkout                    7.8 sec
test_checkout-checkout-noop          0.95 sec
test_checkout-checkout-update       25 sec
---------------------------------------------
```

this PR:
```
------------------------------------------
test_checkout-checkout                  12.2 sec
test_checkout-checkout-noop        2 sec
test_checkout-checkout-update     5.8 sec
------------------------------------------
```

so we've gained better performance in complex cases, but lost performance in simple cases because we've also adopted all performance problems of `stage()` that need to be solved anyway (but they are common for multiple commands, so one optimization round there will improve many), plus diffing big trees is kinda slow in python when using a naive approach (need to polish it out and/or add a simple c extension, e.g. like dulwich). Still, this is moving us in the right direction and allows us to move along with `dvc diff` and `dvc status` and put all problems into one batch (e.g. `stage()` + obj `diff()`), so merging as is (no new release for now though).